requirements.txt
PyYAML
psycopg
hvac
